Remove `as usize` from INDENT_WITH
Somehow, this is no longer required in new rust

+++ src/components/fail_2_ban.rs
@@ -75,7 +75,7 @@ impl Fail2Ban {
                 jail = jail,
                 total = entry.total,
                 current = entry.current,
-                indent = " ".repeat(INDENT_WIDTH as usize),
+                indent = " ".repeat(INDENT_WIDTH),
             );
         }
